implemented sounds for present and mail
