--- SpaceGame/Assets/DoorController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d37d57475dbbebf42814dd67dd4fcbe2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
