--- SpaceGame/Assets/MailSliderController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0885e5100b337444e800765cc199953d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
