--- SpaceGame/Assets/PresentCollisionHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 25758fc2b0919ef409b3de4d4c613cb7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
