+++ SpaceGame/Assets/Scripts/AbstractCollider.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: acdb8ec813284e8f8e8bd8dc9cb987dd
+timeCreated: 1592987395
