+++ SpaceGame/Assets/Scripts/DoorController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4092f5757aa7c5f4ebdcc6cff61d08ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
