+++ SpaceGame/Assets/Scripts/MailSliderController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a53441f796d0634688083949441f784
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
