+++ SpaceGame/Assets/Scripts/PresentCollisionHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a257591be2e3aa047b7ee13d3fba1edd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
